+++ lib/auto_html/filters/image.rb
@@ -1,5 +1,7 @@
-AutoHtml.add_filter(:image) do |text|
-  text.gsub(/http:\/\/.+\.(jpg|jpeg|bmp|gif|png)(\?\S+)?/i) do |match|
-    %|<img src="#{match}" alt=""/>|
+require 'tag_helper'
+
+AutoHtml.add_filter(:image).with({:alt => ''}) do |text, options|
+  text.gsub(/https?:\/\/.+\.(jpg|jpeg|bmp|gif|png)(\?\S+)?/i) do |match|
+    TagHelper.image_tag(match, options)
   end
 end